+++ src/tquality_selenium/elements/__init__.py
@@ -1,7 +1,8 @@
 from tquality_selenium.elements.base_element import BaseElement
 from tquality_selenium.elements.button import Button
+from tquality_selenium.elements.by import By, ByKind
 from tquality_selenium.elements.checkbox import CheckBox
 from tquality_selenium.elements.input import Input
 from tquality_selenium.elements.label import Label
 
-__all__ = ["BaseElement", "Button", "CheckBox", "Input", "Label"]
+__all__ = ["BaseElement", "Button", "By", "ByKind", "CheckBox", "Input", "Label"]